Allow data directories outside pwd
Running ponzu-cms inside a docker container, and maintaining
data files in a volume (or several) was not possible with the
previous directory schema without copying files or binaries
to the volume.

This commit adds several environment variables to allow the
administrator to specify paths for directories while
retaining backwards compatibility.

- PONZU_DATA_DIR : is the directory where all the data and other
                   directories will be stored, still the other
                   directories can be controlled by the other env
                   variables.

- PONZU_TLS_DIR : is the directory to store devcerts and autocerts

- PONZU_ADMINSTATIC_DIR : directory where the admin static files are
                          served from.

- PONZU_UPLOAD_DIR: upload files directory

- PONZU_SEARCH_DIR: index directory for searches